Day 07: Refactored code
