+++ 2024/day07/astar/pq.go
@@ -1,0 +1,35 @@
+package astar
+
+// A PriorityQueue implements heap.Interface and holds Nodes. The
+// PriorityQueue is used to track open nodes by rank.
+type PriorityQueue []*Node
+
+func (pq PriorityQueue) Len() int {
+	return len(pq)
+}
+
+func (pq PriorityQueue) Less(i, j int) bool {
+	return pq[i].rank < pq[j].rank
+}
+
+func (pq PriorityQueue) Swap(i, j int) {
+	pq[i], pq[j] = pq[j], pq[i]
+	pq[i].index = i
+	pq[j].index = j
+}
+
+func (pq *PriorityQueue) Push(x interface{}) {
+	n := len(*pq)
+	no := x.(*Node)
+	no.index = n
+	*pq = append(*pq, no)
+}
+
+func (pq *PriorityQueue) Pop() interface{} {
+	old := *pq
+	n := len(old)
+	no := old[n-1]
+	no.index = -1
+	*pq = old[0 : n-1]
+	return no
+}
